+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/codec/SimulcastVideoEncoderFactoryWrapper.kt
@@ -1,5 +1,6 @@
 package jp.shiguredo.sora.sdk.codec
 
+import jp.shiguredo.sora.sdk.channel.option.SoraVideoOption
 import jp.shiguredo.sora.sdk.util.SoraLogger
 import org.webrtc.EglBase
 import org.webrtc.HardwareVideoEncoderFactory
@@ -9,7 +10,6 @@ import org.webrtc.VideoCodecInfo
 import org.webrtc.VideoCodecStatus
 import org.webrtc.VideoEncoder
 import org.webrtc.VideoEncoderFactory
-import org.webrtc.VideoEncoderFallback
 import org.webrtc.VideoFrame
 import java.util.concurrent.Callable
 import java.util.concurrent.ExecutorService
@@ -18,47 +18,9 @@ import java.util.concurrent.Executors
 internal class SimulcastVideoEncoderFactoryWrapper(
     sharedContext: EglBase.Context?,
     enableIntelVp8Encoder: Boolean,
-    enableH264HighProfile: Boolean
+    enableH264HighProfile: Boolean,
+    videoCodec: SoraVideoOption.Codec
 ) : VideoEncoderFactory {
-
-    /*
-     * ソフトウェアエンコーダーの利用を優先するファクトリーです。
-     * 指定されたコーデックにソフトウェアエンコーダーが対応していない場合にハードウェアエンコーダーを利用します。
-     * ただし、このクラスは libwebrtc の問題を回避するためのワークアラウンドで実質的に用途はありません。
-     *
-     * libwebrtc でサイマルキャストを扱うには SimulcastEncoderAdapter を利用します。
-     * SimulcastEncoderAdapter はプライマリのエンコーダーとフォールバックのエンコーダー
-     * (指定されたコーデックにプライマリのエンコーダーが対応していない場合に利用されます)
-     * を持ちます。プライマリに HardwareVideoEncoderFactory 、
-     * フォールバックに SoftwareVideoEncoderFactory を指定すると、
-     * H.264 の使用時に libwebrtc がクラッシュします。
-     * これは SoftwareVideoEncoderFactory が H.264 に非対応であり、
-     * createEncoder() が null を返すのに対して libwebrtc 側が null 時の処理に対応していないためです。
-     * createEncoder() はフォールバックの利用の有無に関わらず実行されるので、
-     * null を回避するために HardwareVideoEncoderFactory に処理を委譲しています。
-     * プライマリ・フォールバックの両方で HardwareVideoEncoderFactory を使うことになりますが、特に問題ありません。
-     */
-    private class Fallback(private val hardwareVideoEncoderFactory: VideoEncoderFactory) : VideoEncoderFactory {
-
-        private val softwareVideoEncoderFactory: VideoEncoderFactory = SoftwareVideoEncoderFactory()
-
-        override fun createEncoder(info: VideoCodecInfo): VideoEncoder? {
-            val softwareEncoder = softwareVideoEncoderFactory.createEncoder(info)
-            val hardwareEncoder = hardwareVideoEncoderFactory.createEncoder(info)
-            return if (hardwareEncoder != null && softwareEncoder != null) {
-                VideoEncoderFallback(hardwareEncoder, softwareEncoder)
-            } else {
-                softwareEncoder ?: hardwareEncoder
-            }
-        }
-
-        override fun getSupportedCodecs(): Array<VideoCodecInfo> {
-            val supportedCodecInfos: MutableList<VideoCodecInfo> = mutableListOf()
-            supportedCodecInfos.addAll(softwareVideoEncoderFactory.supportedCodecs)
-            supportedCodecInfos.addAll(hardwareVideoEncoderFactory.supportedCodecs)
-            return supportedCodecInfos.toTypedArray()
-        }
-    }
 
     // ストリーム単位のエンコーダをラップした上で以下を行うクラス。
     // - スレッドをひとつ起動する
@@ -81,6 +43,7 @@ internal class SimulcastVideoEncoderFactoryWrapper(
                     SoraLogger.i(
                         TAG,
                         """initEncode() thread=${Thread.currentThread().name} [${Thread.currentThread().id}]
+                |  encoder=${encoder.implementationName}
                 |  streamSettings:
                 |    numberOfCores=${settings.numberOfCores}
                 |    width=${settings.width}
@@ -165,7 +128,7 @@ internal class SimulcastVideoEncoderFactoryWrapper(
     }
 
     private val primary: VideoEncoderFactory
-    private val fallback: VideoEncoderFactory
+    private val fallback: VideoEncoderFactory?
     private val native: SimulcastVideoEncoderFactory
 
     init {
@@ -173,7 +136,14 @@ internal class SimulcastVideoEncoderFactoryWrapper(
             sharedContext, enableIntelVp8Encoder, enableH264HighProfile
         )
         primary = StreamEncoderWrapperFactory(hardwareVideoEncoderFactory)
-        fallback = StreamEncoderWrapperFactory(Fallback(primary))
+
+        // H.264 のサイマルキャストを利用する場合は fallback に null を設定する
+        // Sora Android SDK では SW の H.264 を無効化しているため fallback に設定できるものがない
+        fallback = if (videoCodec != SoraVideoOption.Codec.H264) {
+            SoftwareVideoEncoderFactory()
+        } else {
+            null
+        }
         native = SimulcastVideoEncoderFactory(primary, fallback)
     }